ArmorData 추가 및 데이터베이스 생성

+++ Assets/Resources/Scrips/Data/WeaponData.cs
@@ -26,7 +26,7 @@ public class WeaponDataInfo
     public int armorBreak;
     public int critical;
     public float range;
-    public int MOA;
+    public float MOA;
     public int stability;
     public int rebound;
     [Space(5f)]
