+++ Assets/Resources/Scrips/Weapon.cs
@@ -27,7 +27,7 @@ public class Weapon : MonoBehaviour
     [Tooltip("방어구 손상")] public int armorBreak;
     [Tooltip("파편화")] public int critical;
     [Tooltip("사거리")] public float range;
-    [Tooltip("정확도")] public int MOA;
+    [Tooltip("정확도")] public float MOA;
     [Tooltip("안정성")] public int stability;
     [Tooltip("반동")] public int rebound;
     [Space(5f)]
